making initialize command service callable
